Abstract out by backing source to Repositories
This breaks out whether the Model is backed by a Custom Table
or a WordPress object into individual Repositories responsible
for handling the fetching and persisting the models. This allows
the EntityManager to defer to the repositories for querying and
saving particular classes and allows us to break up the location
of the logic based on backing source.